Module nphysics3df64​::integration​::euler​

Euler integration functions.

Functions

displacement
Computes the transformation matrix required to move an object with a lin_vel linear velocity, a ang_vil angular velocity, and a center of mass center_of_mass, during the time step dt.

explicit_integrate
Explicit Euler integrator.

explicit_integrate_wo_rotation
Explicit Euler integrator. This will not update the rotational components.

semi_implicit_integrate
Semi-implicit Euler integrator.

semi_implicit_integrate_wo_rotation
Semi-implicit Euler integrator. This will not update the rotational components.