Module nphysics3df64​::resolution​

Constraint resolution.

Reexports

pub use resolution::constraint::contact_equation::{VelocityAndPosition, Velocity, VelocityAndPositionThresold};

Structs

AccumulatedImpulseSolver
Constraint solver using the projected gauss seidel algorithm and warm-starting.

ContactIdentifier
The identifier of a contact stored in the impulse cache.

ImpulseCache

Velocities
Structure holding the result of the projected gauss seidel solver.

VelocityConstraint
A constraint of velocity at a point of contact.

Enums

CorrectionMode
The correction coefficient used by the constraint solver.

Traits

Solver
Trait implemented by constraint solvers.

Functions

projected_gauss_seidel_solve
Solve a set of velocity constraints using the projected gauss seidel solver.