Struct nphysics::world::World

pub struct World {
    // some fields omitted
}

The physics world.

This is the main structure of the physics engine.

Methods

impl World

fn new() -> World

Creates a new physics world.

fn step(&mut self, dt: Scalar)

Updates the physics world.

fn add_body(&mut self, rb: RigidBody) -> RigidBodyHandle

Adds a rigid body to the physics world.

fn remove_body(&mut self, b: &RigidBodyHandle)

Remove a rigid body from the physics world.

fn forces_generator(&mut self) -> &mut BodyForceGenerator

Gets a mutable reference to the force generator.

fn integrator(&mut self) -> &mut BodySmpEulerIntegrator

Gets a mutable reference to the position and orientation integrator.

fn collision_world(&mut self) -> &mut RigidBodyCollisionWorld

Gets a mutable reference to the collision detector.

fn ccd_manager(&mut self) -> &mut TranslationalCCDMotionClamping

Gets a mutable reference to the CCD manager.

fn joint_manager(&mut self) -> &mut JointManager

Gets a mutable reference to the joint manager.

fn constraints_solver(&mut self) -> &mut AccumulatedImpulseSolver

Gets a mutable reference to the constraint solver.

fn set_gravity(&mut self, gravity: Vect)

Sets the linear acceleration afecting every dynamic rigid body.

fn set_angular_acceleration(&mut self, accel: Orientation)

Sets the angular acceleration afecting every dynamic rigid body.

fn gravity(&self) -> Vect

Gets the linear acceleration afecting every dynamic rigid body.

fn angular_acceleration(&self) -> Orientation

Gets the angular acceleration afecting every dynamic rigid body.

fn interferences_with_ray<F: FnMut(&RigidBodyHandle, RayIntersection<Scalar, Vect>) -> ()>(&mut self, ray: &Ray<Point, Vect>, f: F)

Gets every body intersected by a given ray.

fn interferences_with_point<F: FnMut(&RigidBodyHandle) -> ()>(&mut self, p: &Point, f: F)

Gets every body that contain a specific point.

fn interferences_with_aabb<F: FnMut(&RigidBodyHandle) -> ()>(&mut self, aabb: &AABB<Point>, f: F)

Gets every body that intersects a specific AABB.

fn add_ccd_to(&mut self, body: &RigidBodyHandle, motion_thresold: Scalar)

Adds continuous collision detection to the given rigid body.

fn add_ball_in_socket(&mut self, joint: BallInSocket) -> Rc<RefCell<BallInSocket>>

Adds a ball-in-socket joint to the world.

fn remove_ball_in_socket(&mut self, joint: &Rc<RefCell<BallInSocket>>)

Removes a ball-in-socket joint from the world.

fn add_fixed(&mut self, joint: Fixed) -> Rc<RefCell<Fixed>>

Adds a fixed joint to the world.

fn remove_fixed(&mut self, joint: &Rc<RefCell<Fixed>>)

Removes a fixed joint from the world.

fn interferences(&mut self, out: &mut Vec<Constraint>)

Collects every interferences detected since the last update.

fn bodies(&self) -> RigidBodies

An iterator visiting all rigid bodies on this world.

fn register_contact_signal_handler<H>(&mut self, name: &str, handler: H) where H: ContactSignalHandler<RigidBodyHandle> + 'static

Registers a handler for contact start/stop events.

fn unregister_contact_signal_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.