Struct nphysics3df64​::world​::World​

pub struct World {
    // some fields omitted
}

The physics world.

This is the main structure of the physics engine.

Methods

impl World

fn new() -> World

Creates a new physics world.

fn step(&mut self, dt: Scalar)

Updates the physics world.

fn add_body(&mut self, b: Rc<RefCell<RigidBody>>)

Adds a rigid body to the physics world.

fn remove_body(&mut self, b: &Rc<RefCell<RigidBody>>)

Remove a rigid body from the physics world.

fn forces_generator<'a>(&'a mut self) -> &'a mut BodyForceGenerator

Gets a mutable reference to the force generator.

fn integrator<'a>(&'a mut self) -> &'a mut BodySmpEulerIntegrator

Gets a mutable reference to the position and orientation integrator.

fn collision_detector<'a>(&'a mut self) -> &'a mut BodiesBodies<WorldBroadPhase>

Gets a mutable reference to the collision detector.

fn broad_phase<'a>(&'a mut self) -> &'a mut WorldBroadPhase

Gets a mutable reference to the broad phase.

fn joint_manager<'a>(&'a mut self) -> &'a mut JointManager

Gets a mutable reference to the joint manager.

fn constraints_solver<'a>(&'a mut self) -> &'a mut AccumulatedImpulseSolver

Gets a mutable reference to the constraint solver.

fn set_gravity(&mut self, gravity: Vect)

Sets the linear acceleration afecting every dynamic rigid body.

fn set_angular_acceleration(&mut self, accel: Orientation)

Sets the angular acceleration afecting every dynamic rigid body.

fn gravity(&self) -> Vect

Gets the linear acceleration afecting every dynamic rigid body.

fn angular_acceleration(&self) -> Orientation

Gets the angular acceleration afecting every dynamic rigid body.

fn cast_ray(&mut self, ray: &Ray, out: &mut Vec<(Rc<RefCell<RigidBody>>, Scalar)>)

Gets every body intersected by a given ray.

fn add_ball_in_socket(&mut self, joint: Rc<RefCell<BallInSocket>>)

Adds a ball-in-socket joint to the world.

fn remove_ball_in_socket(&mut self, joint: &Rc<RefCell<BallInSocket>>)

Removes a ball-in-socket joint from the world.

fn add_fixed(&mut self, joint: Rc<RefCell<Fixed>>)

Adds a fixed joint to the world.

fn remove_fixed(&mut self, joint: &Rc<RefCell<Fixed>>)

Removes a fixed joint from the world.

fn interferences(&mut self, out: &mut Vec<Constraint>)

Collects every interferences detected since the last update.